# Conflicts:
#	src/main/java/org/generation/libri/generationlibrary/controller/BookController.java

+++ src/main/java/org/generation/libri/generationlibrary/controller/admin/CategoryController.java
@@ -1,7 +1,9 @@
-package org.generation.libri.generationlibrary.controller;
+package org.generation.libri.generationlibrary.controller.admin;
 
 import jakarta.validation.Valid;
+import org.generation.libri.generationlibrary.model.Book;
 import org.generation.libri.generationlibrary.model.Category;
+import org.generation.libri.generationlibrary.repository.BookRepository;
 import org.generation.libri.generationlibrary.repository.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -20,8 +22,9 @@ import java.util.Optional;
 public class CategoryController {
     @Autowired
     CategoryRepository categoryRepository;
+    @Autowired
+    private BookRepository bookRepository;
 
-    //Metodo che mostra la lista delle categories
     @GetMapping
     public String index(Model model) {
         model.addAttribute("categoryList", categoryRepository.findAll());
@@ -29,7 +32,6 @@ public class CategoryController {
         return "/admin/categories/list";
     }
 
-    //Metodo che crea una nuova categoria
     @PostMapping("/create")
     public String doCreate(@ModelAttribute("categoryObj") Category categoryform, RedirectAttributes redirectAttributes) {
         categoryRepository.save(categoryform);
@@ -37,7 +39,6 @@ public class CategoryController {
         return "redirect:/admin/categories";
     }
 
-    //Metodo per l'update
     @GetMapping("/update/{catId}")
     public String update(@PathVariable Integer id, Model model) {
         Optional<Category> result = categoryRepository.findById(id);
@@ -58,12 +59,22 @@ public class CategoryController {
         return "redirect:/admin/categories/list";
     }
 
-
-    //Metodo che cancella una categoria
     @PostMapping("/delete/{catId}")
-    public String delete(@PathVariable("catId") int id) {
-        categoryRepository.deleteById(id);
-        return "redirect:/admin/categories";
+    public String deleteSpecialOffer(@PathVariable("catId") Integer id) {
+        Optional<Category> categoryOptional = categoryRepository.findById(id);
+        if (categoryOptional.isPresent()) {
+            Category category = categoryOptional.get();
+            for (Book book : category.getBooks()) {
+                book.getCategories().remove(category);
+                bookRepository.save(book);
+            }
+            category.getBooks().clear();
+            categoryRepository.save(category);
+            categoryRepository.deleteById(id);
+            return "redirect:/admin/categories";
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
 
     //SEARCH
